Moved analyzer where we want to make a fix (and hence add files) into seperate folder. Did not move other analyzers as they will not have additional files. Moved all namespace tests into seperate folders, since they all consist of at least two files, and the analyzer for which we will make fix, will eventually consist of many more.
