+++ Liversen.DependencyCop.Tests/DescendantNamespaceAccess/AnalyzerTestCode.cs
