+++ Liversen.DependencyCop.Tests/Liversen.DependencyCop.Tests.csproj
@@ -5,14 +5,14 @@
     <IsTestProject>true</IsTestProject>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Remove="DescendantNamespaceAccessAnalyzerTestCode.cs" />
-    <Compile Remove="NamespaceCycleAnalyzerTestCode.cs" />
-    <Compile Remove="UsingNamespaceStatementAnalyzerTestCode.cs" />
+    <Compile Remove="DescendantNamespaceAccess\AnalyzerTestCode.cs" />
+    <Compile Remove="NamespaceCycle\AnalyzerTestCode.cs" />
+    <Compile Remove="UsingNamespaceStatement\AnalyzerTestCode.cs" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="DescendantNamespaceAccessAnalyzerTestCode.cs" />
-    <EmbeddedResource Include="NamespaceCycleAnalyzerTestCode.cs" />
-    <EmbeddedResource Include="UsingNamespaceStatementAnalyzerTestCode.cs" />
+    <EmbeddedResource Include="DescendantNamespaceAccess\AnalyzerTestCode.cs" />
+    <EmbeddedResource Include="NamespaceCycle\AnalyzerTestCode.cs" />
+    <EmbeddedResource Include="UsingNamespaceStatement\AnalyzerTestCode.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Analyzer.Testing" Version="1.1.2" />
