+++ Liversen.DependencyCop.Tests/NamespaceCycle/AnalyzerTestCode.cs
