+++ Liversen.DependencyCop.Tests/UsingNamespaceStatement/AnalyzerTest.cs
@@ -8,9 +8,9 @@ using Shouldly;
 using Xunit;
 #pragma warning disable CS0618 // Type or member is obsolete
 
-namespace Liversen.DependencyCop
+namespace Liversen.DependencyCop.UsingNamespaceStatement
 {
-    public class UsingNamespaceStatementAnalyzerTest
+    public class AnalyzerTest
     {
         const string DisallowedNamespacePrefixesString = "UsingNamespaceStatementAnalyzer";
 
@@ -26,7 +26,7 @@ namespace Liversen.DependencyCop
             var expected = new DiagnosticResult("DC1001", DiagnosticSeverity.Warning)
                 .WithLocation(1, 1)
                 .WithMessage("Do not use 'UsingNamespaceStatementAnalyzer.Account' in a using statement, use fully-qualified names");
-            var test = new CSharpAnalyzerTest<UsingNamespaceStatementAnalyzer, XUnitVerifier>
+            var test = new CSharpAnalyzerTest<Analyzer, XUnitVerifier>
             {
                 TestState =
                 {
@@ -45,7 +45,7 @@ namespace Liversen.DependencyCop
             var code = EmbeddedResourceHelpers.GetFromCallingAssembly($"{GetType().FullName}Code.cs");
             var expected = new DiagnosticResult("DC1004", DiagnosticSeverity.Warning)
                 .WithMessage("A list of disallowed namespaces must be configured for rule DC1001");
-            var test = new CSharpAnalyzerTest<UsingNamespaceStatementAnalyzer, XUnitVerifier>
+            var test = new CSharpAnalyzerTest<Analyzer, XUnitVerifier>
             {
                 TestState =
                 {
