+++ Liversen.DependencyCop/UsingNamespaceStatement/Analyzer.cs
@@ -6,10 +6,10 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 
-namespace Liversen.DependencyCop
+namespace Liversen.DependencyCop.UsingNamespaceStatement
 {
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
-    public class UsingNamespaceStatementAnalyzer : DiagnosticAnalyzer
+    public class Analyzer : DiagnosticAnalyzer
     {
         public const string RuleDC1001Id = "DC1001";
         const string DotnetDiagnosticOptionName = "dotnet_diagnostic.DC1001_NamespacePrefixes";
